requirements.txt
pytest==8.0.0
selene==2.0.0rc7
Appium-Python-Client==3.1.1
allure-pytest==2.13.2
python-dotenv==1.0.1
pydantic==2.6.0
requests==2.31.0